+++ src/TugasBesar.Core/Controllers/Interfaces/IAkunPegawaiApi.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using TugasBesar.Core.DTO.Request;
 using TugasBesar.Core.DTO.Response;
-using TugasBesar.Core.Models;
 
 namespace TugasBesar.Core.Controllers.Interfaces
 {
@@ -16,9 +15,9 @@ namespace TugasBesar.Core.Controllers.Interfaces
         Task TambahAkun([Body] AkunRequestDTO request);
 
         [Put("/api/AkunPegawai/{id}")]
-        Task UpdateAkun(int id, [Body] AkunRequestDTO request);
+        Task UpdateAkun(long id, [Body] AkunRequestDTO request);
 
         [Delete("/api/AkunPegawai/{id}")]
-        Task HapusAkun(int id);
+        Task HapusAkun(long id);
     }
 }
