+++ src/TugasBesar.Core/DTO/Response/AkunResponseDTO.cs
@@ -4,9 +4,12 @@ using System.Text;
 
 namespace TugasBesar.Core.DTO.Response
 {
+    /// <summary>
+    /// Data Transfer Object untuk mengirimkan data akun pegawai ke client secara aman.
+    /// </summary>
     public class AkunResponseDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Role { get; set; } = string.Empty;
     }
